fix: reduce schematics test time
