+++ feature-libs/pickup-in-store/schematics/add-pickup-in-store/index_spec.ts
@@ -5,19 +5,14 @@ import {
   UnitTestTree,
 } from '@angular-devkit/schematics/testing';
 import {
-  Schema as ApplicationOptions,
-  Style,
-} from '@schematics/angular/application/schema';
-import { Schema as WorkspaceOptions } from '@schematics/angular/workspace/schema';
-import {
+  cartBaseFeatureModulePath,
+  generateDefaultWorkspace,
+  LibraryOptions as SpartacusPickupInStoreOptions,
+  orderFeatureModulePath,
   PICKUP_IN_STORE_FEATURE_NAME,
+  pickupInStoreFeatureModulePath,
   SPARTACUS_PICKUP_IN_STORE,
   SPARTACUS_SCHEMATICS,
-  SpartacusOptions,
-  LibraryOptions as SpartacusPickupInStoreOptions,
-  cartBaseFeatureModulePath,
-  orderFeatureModulePath,
-  pickupInStoreFeatureModulePath,
   storeFinderFeatureModulePath,
   userFeatureModulePath,
 } from '@spartacus/schematics';
@@ -35,27 +30,6 @@ describe('Spartacus Pickup in Store schematics: ng-add', () => {
 
   let appTree: UnitTestTree;
 
-  const workspaceOptions: WorkspaceOptions = {
-    name: 'workspace',
-    version: '0.5.0',
-  };
-
-  const appOptions: ApplicationOptions = {
-    name: 'schematics-test',
-    inlineStyle: false,
-    inlineTemplate: false,
-    style: Style.Scss,
-    skipTests: false,
-    projectRoot: '',
-    standalone: false,
-  };
-
-  const spartacusDefaultOptions: SpartacusOptions = {
-    project: 'schematics-test',
-    lazy: true,
-    features: [],
-  };
-
   const libraryNoFeaturesOptions: SpartacusPickupInStoreOptions = {
     project: 'schematics-test',
     lazy: true,
@@ -67,35 +41,9 @@ describe('Spartacus Pickup in Store schematics: ng-add', () => {
     features: [PICKUP_IN_STORE_FEATURE_NAME],
   };
 
-  beforeEach(async () => {
-    schematicRunner.registerCollection(
-      SPARTACUS_SCHEMATICS,
-      '../../projects/schematics/src/collection.json'
-    );
-
-    appTree = await schematicRunner.runExternalSchematic(
-      '@schematics/angular',
-      'workspace',
-      workspaceOptions
-    );
-
-    appTree = await schematicRunner.runExternalSchematic(
-      '@schematics/angular',
-      'application',
-      appOptions,
-      appTree
-    );
-
-    appTree = await schematicRunner.runExternalSchematic(
-      SPARTACUS_SCHEMATICS,
-      'ng-add',
-      { ...spartacusDefaultOptions, name: 'schematics-test' },
-      appTree
-    );
-  });
-
   describe('Without features', () => {
-    beforeEach(async () => {
+    beforeAll(async () => {
+      appTree = await generateDefaultWorkspace(schematicRunner, appTree);
       appTree = await schematicRunner.runSchematic(
         'ng-add',
         { ...libraryNoFeaturesOptions, features: [] },
@@ -135,7 +83,8 @@ describe('Spartacus Pickup in Store schematics: ng-add', () => {
 
   describe('Pick Up In Store feature', () => {
     describe('general setup', () => {
-      beforeEach(async () => {
+      beforeAll(async () => {
+        appTree = await generateDefaultWorkspace(schematicRunner, appTree);
         appTree = await schematicRunner.runSchematic(
           'ng-add',
           pickupInStoreFeatureOptions,
@@ -177,7 +126,8 @@ describe('Spartacus Pickup in Store schematics: ng-add', () => {
     });
 
     describe('eager loading', () => {
-      beforeEach(async () => {
+      beforeAll(async () => {
+        appTree = await generateDefaultWorkspace(schematicRunner, appTree);
         appTree = await schematicRunner.runSchematic(
           'ng-add',
           { ...pickupInStoreFeatureOptions, lazy: false },
