+++ feature-libs/product-multi-dimensional/schematics/add-product-multi-dimensional/index_spec.ts
@@ -5,19 +5,14 @@ import {
   UnitTestTree,
 } from '@angular-devkit/schematics/testing';
 import {
-  Schema as ApplicationOptions,
-  Style,
-} from '@schematics/angular/application/schema';
-import { Schema as WorkspaceOptions } from '@schematics/angular/workspace/schema';
-import {
+  generateDefaultWorkspace,
+  LibraryOptions as SpartacusProductOptions,
   PRODUCT_MULTI_DIMENSIONAL_LIST_FEATURE_NAME,
   PRODUCT_MULTI_DIMENSIONAL_SELECTOR_FEATURE_NAME,
-  SPARTACUS_PRODUCT,
-  SPARTACUS_SCHEMATICS,
-  SpartacusOptions,
-  LibraryOptions as SpartacusProductOptions,
   productMultiDimensionalListFeatureModulePath,
   productMultiDimensionalSelectorFeatureModulePath,
+  SPARTACUS_PRODUCT,
+  SPARTACUS_SCHEMATICS,
 } from '@spartacus/schematics';
 import * as path from 'path';
 import { peerDependencies } from '../../package.json';
@@ -32,27 +27,6 @@ describe('Spartacus Product Multi-Dimensional schematics: ng-add', () => {
   );
 
   let appTree: UnitTestTree;
-
-  const workspaceOptions: WorkspaceOptions = {
-    name: 'workspace',
-    version: '0.5.0',
-  };
-
-  const appOptions: ApplicationOptions = {
-    name: 'schematics-test',
-    inlineStyle: false,
-    inlineTemplate: false,
-    style: Style.Scss,
-    skipTests: false,
-    projectRoot: '',
-    standalone: false,
-  };
-
-  const spartacusDefaultOptions: SpartacusOptions = {
-    project: 'schematics-test',
-    lazy: true,
-    features: [],
-  };
 
   const libraryNoFeaturesOptions: SpartacusProductOptions = {
     project: 'schematics-test',
@@ -70,35 +44,9 @@ describe('Spartacus Product Multi-Dimensional schematics: ng-add', () => {
     features: [PRODUCT_MULTI_DIMENSIONAL_LIST_FEATURE_NAME],
   };
 
-  beforeEach(async () => {
-    schematicRunner.registerCollection(
-      SPARTACUS_SCHEMATICS,
-      '../../projects/schematics/src/collection.json'
-    );
-
-    appTree = await schematicRunner.runExternalSchematic(
-      '@schematics/angular',
-      'workspace',
-      workspaceOptions
-    );
-
-    appTree = await schematicRunner.runExternalSchematic(
-      '@schematics/angular',
-      'application',
-      appOptions,
-      appTree
-    );
-
-    appTree = await schematicRunner.runExternalSchematic(
-      SPARTACUS_SCHEMATICS,
-      'ng-add',
-      { ...spartacusDefaultOptions, name: 'schematics-test' },
-      appTree
-    );
-  });
-
   describe('Without features', () => {
     beforeEach(async () => {
+      appTree = await generateDefaultWorkspace(schematicRunner, appTree);
       appTree = await schematicRunner.runSchematic(
         'ng-add',
         libraryNoFeaturesOptions,
@@ -140,7 +88,8 @@ describe('Spartacus Product Multi-Dimensional schematics: ng-add', () => {
 
   describe('selector feature', () => {
     describe('general setup', () => {
-      beforeEach(async () => {
+      beforeAll(async () => {
+        appTree = await generateDefaultWorkspace(schematicRunner, appTree);
         appTree = await schematicRunner.runSchematic(
           'ng-add',
           multiDimensionalSelectorOptions,
@@ -169,7 +118,8 @@ describe('Spartacus Product Multi-Dimensional schematics: ng-add', () => {
     });
 
     describe('eager loading', () => {
-      beforeEach(async () => {
+      beforeAll(async () => {
+        appTree = await generateDefaultWorkspace(schematicRunner, appTree);
         appTree = await schematicRunner.runSchematic(
           'ng-add',
           { ...multiDimensionalSelectorOptions, lazy: false },
@@ -188,7 +138,8 @@ describe('Spartacus Product Multi-Dimensional schematics: ng-add', () => {
 
   describe('list feature', () => {
     describe('general setup', () => {
-      beforeEach(async () => {
+      beforeAll(async () => {
+        appTree = await generateDefaultWorkspace(schematicRunner, appTree);
         appTree = await schematicRunner.runSchematic(
           'ng-add',
           multiDimensionalListOptions,
@@ -205,7 +156,8 @@ describe('Spartacus Product Multi-Dimensional schematics: ng-add', () => {
     });
 
     describe('eager loading', () => {
-      beforeEach(async () => {
+      beforeAll(async () => {
+        appTree = await generateDefaultWorkspace(schematicRunner, appTree);
         appTree = await schematicRunner.runSchematic(
           'ng-add',
           { ...multiDimensionalListOptions, lazy: false },
