+++ feature-libs/requested-delivery-date/schematics/add-requested-delivery-date/index_spec.ts
@@ -5,17 +5,12 @@ import {
   UnitTestTree,
 } from '@angular-devkit/schematics/testing';
 import {
-  Schema as ApplicationOptions,
-  Style,
-} from '@schematics/angular/application/schema';
-import { Schema as WorkspaceOptions } from '@schematics/angular/workspace/schema';
-import {
-  REQUESTED_DELIVERY_DATE_FEATURE_NAME,
+  generateDefaultWorkspace,
   LibraryOptions as RequestedDeliveryDateOptions,
+  REQUESTED_DELIVERY_DATE_FEATURE_NAME,
+  requestedDeliveryDateFeatureModulePath,
   SPARTACUS_REQUESTED_DELIVERY_DATE,
   SPARTACUS_SCHEMATICS,
-  SpartacusOptions,
-  requestedDeliveryDateFeatureModulePath,
 } from '@spartacus/schematics';
 import * as path from 'path';
 import { peerDependencies } from '../../package.json';
@@ -31,27 +26,6 @@ describe('Spartacus Requested Delivery Date schematics: ng-add', () => {
 
   let appTree: UnitTestTree;
 
-  const workspaceOptions: WorkspaceOptions = {
-    name: 'workspace',
-    version: '0.5.0',
-  };
-
-  const appOptions: ApplicationOptions = {
-    name: 'schematics-test',
-    inlineStyle: false,
-    inlineTemplate: false,
-    style: Style.Scss,
-    skipTests: false,
-    projectRoot: '',
-    standalone: false,
-  };
-
-  const spartacusDefaultOptions: SpartacusOptions = {
-    project: 'schematics-test',
-    lazy: true,
-    features: [],
-  };
-
   const libraryNoFeaturesOptions: RequestedDeliveryDateOptions = {
     project: 'schematics-test',
     lazy: true,
@@ -63,35 +37,9 @@ describe('Spartacus Requested Delivery Date schematics: ng-add', () => {
     features: [REQUESTED_DELIVERY_DATE_FEATURE_NAME],
   };
 
-  beforeEach(async () => {
-    schematicRunner.registerCollection(
-      SPARTACUS_SCHEMATICS,
-      '../../projects/schematics/src/collection.json'
-    );
-
-    appTree = await schematicRunner.runExternalSchematic(
-      '@schematics/angular',
-      'workspace',
-      workspaceOptions
-    );
-
-    appTree = await schematicRunner.runExternalSchematic(
-      '@schematics/angular',
-      'application',
-      appOptions,
-      appTree
-    );
-
-    appTree = await schematicRunner.runExternalSchematic(
-      SPARTACUS_SCHEMATICS,
-      'ng-add',
-      { ...spartacusDefaultOptions, name: 'schematics-test' },
-      appTree
-    );
-  });
-
   describe('Without features', () => {
-    beforeEach(async () => {
+    beforeAll(async () => {
+      appTree = await generateDefaultWorkspace(schematicRunner, appTree);
       appTree = await schematicRunner.runSchematic(
         'ng-add',
         libraryNoFeaturesOptions,
@@ -133,7 +81,8 @@ describe('Spartacus Requested Delivery Date schematics: ng-add', () => {
 
   describe('Requested Delivery Date feature', () => {
     describe('general setup', () => {
-      beforeEach(async () => {
+      beforeAll(async () => {
+        appTree = await generateDefaultWorkspace(schematicRunner, appTree);
         appTree = await schematicRunner.runSchematic(
           'ng-add',
           rddOptions,
@@ -162,7 +111,8 @@ describe('Spartacus Requested Delivery Date schematics: ng-add', () => {
     });
 
     describe('eager loading', () => {
-      beforeEach(async () => {
+      beforeAll(async () => {
+        appTree = await generateDefaultWorkspace(schematicRunner, appTree);
         appTree = await schematicRunner.runSchematic(
           'ng-add',
           { ...rddOptions, lazy: false },
