+++ integration-libs/cdc/schematics/add-cdc/index_spec.ts
@@ -4,31 +4,27 @@ import {
   UnitTestTree,
 } from '@angular-devkit/schematics/testing';
 import {
-  Schema as ApplicationOptions,
-  Style,
-} from '@schematics/angular/application/schema';
-import { Schema as WorkspaceOptions } from '@schematics/angular/workspace/schema';
-import {
   CDC_B2B_FEATURE_NAME,
   CDC_FEATURE_NAME,
+  cdcFeatureModulePath,
+  generateDefaultWorkspace,
+  LibraryOptions as SpartacusCdcOptions,
   ORGANIZATION_ADMINISTRATION_FEATURE_NAME,
   ORGANIZATION_USER_REGISTRATION_FEATURE_NAME,
+  organizationAdministrationWrapperModulePath,
+  organizationUserRegistrationWrapperModulePath,
   SPARTACUS_ASM,
   SPARTACUS_CDC,
   SPARTACUS_SCHEMATICS,
   SPARTACUS_USER,
-  LibraryOptions as SpartacusCdcOptions,
-  SpartacusOptions,
   USER_ACCOUNT_FEATURE_NAME,
   USER_PROFILE_FEATURE_NAME,
-  cdcFeatureModulePath,
-  organizationAdministrationWrapperModulePath,
-  organizationUserRegistrationWrapperModulePath,
   userAccountWrapperModulePath,
   userProfileWrapperModulePath,
 } from '@spartacus/schematics';
 import * as path from 'path';
 import { peerDependencies } from '../../package.json';
+
 const collectionPath = path.join(__dirname, '../collection.json');
 describe('Spartacus CDC schematics: ng-add', () => {
   const schematicRunner = new SchematicTestRunner(
@@ -36,24 +32,6 @@ describe('Spartacus CDC schematics: ng-add', () => {
     collectionPath
   );
   let appTree: UnitTestTree;
-  const workspaceOptions: WorkspaceOptions = {
-    name: 'workspace',
-    version: '0.5.0',
-  };
-  const appOptions: ApplicationOptions = {
-    name: 'schematics-test',
-    inlineStyle: false,
-    inlineTemplate: false,
-    style: Style.Scss,
-    skipTests: false,
-    projectRoot: '',
-    standalone: false,
-  };
-  const spartacusDefaultOptions: SpartacusOptions = {
-    project: 'schematics-test',
-    lazy: true,
-    features: [],
-  };
   const libraryNoFeaturesOptions: SpartacusCdcOptions = {
     project: 'schematics-test',
     lazy: true,
@@ -67,14 +45,8 @@ describe('Spartacus CDC schematics: ng-add', () => {
     ...libraryNoFeaturesOptions,
     features: [CDC_B2B_FEATURE_NAME],
   };
-  beforeEach(async () => {
-    schematicRunner.registerCollection(
-      SPARTACUS_SCHEMATICS,
-      path.join(
-        __dirname,
-        '../../../../projects/schematics/src/collection.json'
-      )
-    );
+
+  async function generateWorkspace() {
     schematicRunner.registerCollection(
       SPARTACUS_ASM,
       require.resolve('../../../../feature-libs/asm/schematics/collection.json')
@@ -86,26 +58,12 @@ describe('Spartacus CDC schematics: ng-add', () => {
         '../../../../feature-libs/user/schematics/collection.json'
       )
     );
-    appTree = await schematicRunner.runExternalSchematic(
-      '@schematics/angular',
-      'workspace',
-      workspaceOptions
-    );
-    appTree = await schematicRunner.runExternalSchematic(
-      '@schematics/angular',
-      'application',
-      appOptions,
-      appTree
-    );
-    appTree = await schematicRunner.runExternalSchematic(
-      SPARTACUS_SCHEMATICS,
-      'ng-add',
-      { ...spartacusDefaultOptions, name: 'schematics-test' },
-      appTree
-    );
-  });
+    return (appTree = await generateDefaultWorkspace(schematicRunner, appTree));
+  }
+
   describe('Without features', () => {
-    beforeEach(async () => {
+    beforeAll(async () => {
+      appTree = await generateWorkspace();
       appTree = await schematicRunner.runSchematic(
         'ng-add',
         libraryNoFeaturesOptions,
@@ -118,7 +76,8 @@ describe('Spartacus CDC schematics: ng-add', () => {
   });
   describe('CDC-B2C feature', () => {
     describe('validation of jsSDKUrl', () => {
-      beforeEach(async () => {
+      beforeAll(async () => {
+        appTree = await generateWorkspace();
         appTree = await schematicRunner.runExternalSchematic(
           SPARTACUS_SCHEMATICS,
           'ng-add',
@@ -140,7 +99,8 @@ describe('Spartacus CDC schematics: ng-add', () => {
       });
     });
     describe('general setup', () => {
-      beforeEach(async () => {
+      beforeAll(async () => {
+        appTree = await generateWorkspace();
         appTree = await schematicRunner.runExternalSchematic(
           SPARTACUS_SCHEMATICS,
           'ng-add',
@@ -195,7 +155,8 @@ describe('Spartacus CDC schematics: ng-add', () => {
       });
     });
     describe('eager loading', () => {
-      beforeEach(async () => {
+      beforeAll(async () => {
+        appTree = await generateWorkspace();
         appTree = await schematicRunner.runExternalSchematic(
           SPARTACUS_SCHEMATICS,
           'ng-add',
@@ -219,7 +180,8 @@ describe('Spartacus CDC schematics: ng-add', () => {
   });
   describe('CDC-B2B feature', () => {
     describe('validation of jsSDKUrl', () => {
-      beforeEach(async () => {
+      beforeAll(async () => {
+        appTree = await generateWorkspace();
         appTree = await schematicRunner.runExternalSchematic(
           SPARTACUS_SCHEMATICS,
           'ng-add',
@@ -249,7 +211,8 @@ describe('Spartacus CDC schematics: ng-add', () => {
       });
     });
     describe('general setup', () => {
-      beforeEach(async () => {
+      beforeAll(async () => {
+        appTree = await generateWorkspace();
         appTree = await schematicRunner.runExternalSchematic(
           SPARTACUS_SCHEMATICS,
           'ng-add',
@@ -317,7 +280,8 @@ describe('Spartacus CDC schematics: ng-add', () => {
       });
     });
     describe('eager loading', () => {
-      beforeEach(async () => {
+      beforeAll(async () => {
+        appTree = await generateWorkspace();
         appTree = await schematicRunner.runExternalSchematic(
           SPARTACUS_SCHEMATICS,
           'ng-add',
