+++ projects/schematics/src/shared/utils/index.ts
@@ -20,3 +20,4 @@ export * from './project-tsconfig-paths';
 export * from './schematics-config-utils';
 export * from './test-utils';
 export * from './workspace-utils';
+export * from './generate-default-workspace';
