+++ projects/schematics/src/shared/utils/new-module-utils_spec.ts
@@ -58,7 +58,7 @@ describe('New Module utils', () => {
     lazy: true,
   };
 
-  beforeEach(async () => {
+  beforeAll(async () => {
     tree = await schematicRunner.runExternalSchematic(
       '@schematics/angular',
       'workspace',
