+++ projects/schematics/src/shared/utils/package-utils_spec.ts
@@ -42,7 +42,7 @@ describe('Package utils', () => {
     project: 'schematics-test',
   };
 
-  beforeEach(async () => {
+  beforeAll(async () => {
     appTree = await schematicRunner.runExternalSchematic(
       '@schematics/angular',
       'workspace',
